Can't roll with standard on this one
Turns out we actually depend on a difference between url.parse() and
new url.URL(), in how it handles git SSH/SCP-style urls.

And the old version of standard is too out of date and raises too many
audit warnings.

We can revisit this by perhaps refactoring our use of url.parse so that
it is compatible with new url.URL().