fix: Make file: URLs (mostly) RFC 8909 compliant
The change made in e8c4301a23afb4e39778b4cb048631bf729cb38b (#53) was
reasonable, but had the unfortunate side effect of breaking backwards
compatibility for urls like 'file://.'.

Further investigation showed that we really weren't complying with RFC
8909's specifications regarding file: URLs, and even further
investigation than that showed that it'll be a breaking change if we do.

This patch isolates the non-8909 compliance bits into a single block
that can be disabled with an environment variable and easily removed
when we are willing to take on that breaking change.  The bug fixed by
the commit that broke file://. is still fixed, and we obey RFC8909
otherwise.

PR-URL: https://github.com/npm/npm-package-arg/pull/60
Credit: @isaacs
Close: #60
Reviewed-by: @ruyadorno